+++ NEWS.md
@@ -4,6 +4,12 @@ elevatr 1.0.0 (2024-0x-xx)
 - add argument for specifying temp directory for download files.  Allows users to specify a specific location. (Thanks, @andrew-caudillo: https://github.com/jhollist/elevatr/issues/95)
 - exposed ncpu argument so user can control.  Defaults to 2 if more than 2 cores available.  Thanks to @courtiol for finding this issue and the suggestion!
 
+# Added Functionality
+- Add new `get_elev_profile()` function (#93, @elipousson)
+
+# Fixes
+- Fix issue with `SpatVector` input objects (#91, @elipousson)
+
 elevatr 0.99.0 (2023-09-11)
 =============